+++ src/components/Auth/Signup.js
@@ -1,6 +1,4 @@
 import React, {Component} from 'react';
-import { AppRegistry, StyleSheet, Text, View} from 'react-native';
-import { connect, Provider } from 'react-redux';
 import {Actions} from 'react-native-router-flux';
 import Header from './authHeader';
 
@@ -12,7 +10,7 @@ import { Container, Title, Content, Footer, FooterTab, Button, Left, Right, Body
 //   return {state: state};
 // };
 
-export default connect()(function Signup (state) {
+export default function Signup (state) {
   return (
     <Container>
       <Header />
@@ -25,4 +23,4 @@ export default connect()(function Signup (state) {
     </Content>
     </Container>
   );
-});
+}
