+++ src/components/GroupView.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
-import { AppRegistry, StyleSheet, Text, View} from 'react-native';
-import { connect, Provider } from 'react-redux';
+import { Text, View} from 'react-native';
+import { connect } from 'react-redux';
 import Header from './Header';
 import { Container, Title, Content, Footer, FooterTab, Button, Left, Right, Body, Icon, H1, Card, CardItem, Image } from 'native-base';
 
@@ -19,7 +19,7 @@ const styles = {
 export default function EventView (state) {
   return (
     <Container>
-      <Header title={'name'}/>
+      <Header />
       <Text>GROUP VIEW</Text>
 
     </Container>
